adapts product
