+++ services/catalog/tests/unit/conftest.py
@@ -23,7 +23,6 @@ pytest_plugins = [
     "pytest_simcore.docker_registry",
     "pytest_simcore.docker_swarm",
     "pytest_simcore.environment_configs",
-    "pytest_simcore.faker_products_data",
     "pytest_simcore.faker_users_data",
     "pytest_simcore.postgres_service",
     "pytest_simcore.pydantic_models",
