+++ services/catalog/tests/unit/with_dbs/test_api_routes_services_access_rights.py
@@ -13,6 +13,7 @@ from typing import Any
 from models_library.api_schemas_catalog.service_access_rights import (
     ServiceAccessRightsGet,
 )
+from models_library.products import ProductName
 from pydantic import parse_obj_as
 from respx.router import MockRouter
 from starlette.testclient import TestClient
@@ -29,13 +30,12 @@ pytest_simcore_ops_services_selection = [
 async def test_get_service_access_rights(
     mock_catalog_background_task: None,
     mocked_director_service_api: MockRouter,
-    client: TestClient,
     user: dict[str, Any],
-    products_names: list[str],
+    target_product: ProductName,
     service_catalog_faker: Callable,
     services_db_tables_injector: Callable,
+    client: TestClient,
 ):
-    target_product = products_names[0]  # osparc
     user_id = user["id"]
     user_primary_gid = user["primary_gid"]
 
@@ -76,14 +76,13 @@ async def test_get_service_access_rights(
 async def test_get_service_access_rights_with_more_gids(
     mock_catalog_background_task: None,
     mocked_director_service_api: MockRouter,
-    client: TestClient,
     user: dict[str, Any],
-    products_names: list[str],
+    other_product: ProductName,
     service_catalog_faker: Callable,
     services_db_tables_injector: Callable,
     user_groups_ids: list[int],
+    client: TestClient,
 ):
-    target_product = products_names[1]  # s4l
     user_id = user["id"]
     user_primary_gid = user["primary_gid"]
     everyone_gid, user_gid, team_gid = user_groups_ids
@@ -93,7 +92,7 @@ async def test_get_service_access_rights_with_more_gids(
         "1.0.1",
         team_access="x",
         everyone_access="x",
-        product=target_product,
+        product=other_product,
     )
     # injects fake data in db
     await services_db_tables_injector([fake_service])
@@ -104,7 +103,7 @@ async def test_get_service_access_rights_with_more_gids(
     ).with_query({"user_id": user_id})
     response = client.get(
         f"{url}",
-        headers={"x-simcore-products-name": target_product},
+        headers={"x-simcore-products-name": other_product},
     )
     assert response.status_code == 200
     data = parse_obj_as(ServiceAccessRightsGet, response.json())
